:fire: Make another Angular component to extend on the generic. Seems like a niche minority way of doing things https://stackoverflow.com/questions/57970640/angular-8-generic-component
