+++ src/app/folder/folder.page.html
@@ -18,5 +18,5 @@
     <strong class="capitalize">{{ folder }}</strong>
     <p>Explore <a target="_blank" rel="noopener noreferrer" href="https://ionicframework.com/docs/components">UI Components</a></p>
   </div>
-  <app-autocomplete></app-autocomplete>
+  <app-autocomplete #appAutocomplete passFormControl="this.formControl"></app-autocomplete>
 </ion-content>
